Use modern ACL.as2() to run as SYSTEM
Sometime when invoking the stop function, the management worker thread
would die:

  A thread (Gearman worker 172.17.0.1_manager/563) died unexpectedly due
  to an uncaught exception, this may leave your Jenkins in a bad way and
  is usually indicative of a bug in the code.
  java.lang.StackOverflowError
    at org.acegisecurity.context.SecurityContext$1.getAuthentication(SecurityContext.java:46)
    at org.acegisecurity.context.SecurityContext$2.getAuthentication(SecurityContext.java:60)
    at org.acegisecurity.context.SecurityContext$1.getAuthentication(SecurityContext.java:46)

That started happening after we upgraded to the LTS 2.277 series and I
thus highly suspect it is related to JEP-227:
https://github.com/jenkinsci/jep/blob/master/jep/227/README.adoc

Replace impersonate calls with the modern ACL.as2() introduced in
Jenkins 2.266.

Bump Jenkins requirement to current LTS series (2.277).

Bug: https://phabricator.wikimedia.org/T281737