GearmanProxy: avoid NPE if a finished build has no executor
When the build is finalized we retriever the executor/computer that ran
it to get other works to check the availability lock. Spotbugs
complains `Run.getExecutor()` might be null which really should never
happen.  If that is ever the case, throw an IllegalStateException, it
will be caught and reported by Jenkins runListener.